Add file via upload
Baby/Giant step algorithm, version 2

The algorithm is based on a space–time trade-off. It is a fairly simple modification of trial multiplication, the naive method of finding discrete logarithms.
Given a cyclic group  G of order q, a generator g  of the group and a group element h , the problem is to find an integer x such that
h=g^x
The baby-step giant-step algorithm is based on rewriting x:
x = it + j
t = Ceiling[q^1/2]
 0 < i < m
0 < j < m
Therefore, we have:
g^x=h
g ^{it+j}=h
g^j = h (g^ {- ti} )
The algorithm pre-computes g^j for several values of  j. Then it fixes a t and tries values of  i in the right-hand side of the congruence above, in the manner of trial multiplication. It tests to see if the congruence is satisfied for any value of  j, using the pre-computed values of g^j.